Struct rusty_data::loader::Loader

pub struct Loader<'a> {
    pub has_header: bool,
    pub delimiter: char,
    // some fields omitted
}

Loader struct

Used to load and process data files into tables.

Fields

has_header

True if there are headers present in the file

delimiter

The delimiter character

Methods

impl<'a> Loader<'a>

fn new(has_header: bool, file: &str, delimiter: char) -> Loader

Constructs a new Loader.

fn load_file(self) -> Result<DataTable, Error>

Load the file from the loader with given delimiter.

Pretty rudimentary with poor error handling.

Panics

The input data is not a float.

Failures

The input data is malformed (missing data, non-uniform rows etc.)